requirements.txt
django==1.11
djangorestframework==3.6.4
django-webpack-loader==0.5.0
django-debug-toolbar==1.8